+++ src/main/kotlin/app/revanced/utils/patcher/Patcher.kt
@@ -3,6 +3,7 @@ package app.revanced.utils.patcher
 import app.revanced.cli.command.MainCommand
 import app.revanced.cli.command.MainCommand.args
 import app.revanced.cli.command.MainCommand.logger
+import app.revanced.cli.command.PatchList
 import app.revanced.patcher.Patcher
 import app.revanced.patcher.data.Context
 import app.revanced.patcher.extensions.PatchExtensions.compatiblePackages
@@ -10,7 +11,7 @@ import app.revanced.patcher.extensions.PatchExtensions.include
 import app.revanced.patcher.extensions.PatchExtensions.patchName
 import app.revanced.patcher.patch.Patch
 
-fun Patcher.addPatchesFiltered(allPatches: List<Class<out Patch<Context>>>) {
+fun Patcher.addPatchesFiltered(allPatches: PatchList) {
     val packageName = this.context.packageMetadata.packageName
     val packageVersion = this.context.packageMetadata.packageVersion
 
@@ -18,20 +19,11 @@ fun Patcher.addPatchesFiltered(allPatches: List<Class<out Patch<Context>>>) {
     allPatches.forEach patchLoop@{ patch ->
         val compatiblePackages = patch.compatiblePackages
         val patchName = patch.patchName
-
-        val prefix = "Skipping $patchName, reason"
-
         val args = MainCommand.args.patchArgs?.patchingArgs!!
 
-        if (args.excludedPatches.contains(patchName)) {
-            logger.info("$prefix: manually excluded")
-            return@patchLoop
-        } else if ((!patch.include || args.defaultExclude) && !args.includedPatches.contains(patchName)) {
-            logger.info("$prefix: excluded by default")
-            return@patchLoop
-        }
+        val prefix = "Skipping $patchName"
 
-        if (compatiblePackages == null) logger.warn("$prefix: Missing compatibility annotation. Continuing.")
+        if (compatiblePackages == null) logger.trace("$patchName: No constraint on packages.")
         else {
             if (!compatiblePackages.any { it.name == packageName }) {
                 logger.trace("$prefix: Incompatible with $packageName. This patch is only compatible with ${
@@ -46,9 +38,17 @@ fun Patcher.addPatchesFiltered(allPatches: List<Class<out Patch<Context>>>) {
                 val compatibleWith = compatiblePackages.joinToString(";") { _package ->
                     "${_package.name}: ${_package.versions.joinToString(", ")}"
                 }
-                logger.warn("$prefix: Incompatible with version $packageVersion. This patch is only compatible with version $compatibleWith")
+                logger.warn("$prefix: Incompatible with version $packageVersion. This patch is only compatible with $compatibleWith")
                 return@patchLoop
             }
+        }
+
+        if (args.excludedPatches.contains(patchName)) {
+            logger.info("$prefix: Manually excluded")
+            return@patchLoop
+        } else if ((!patch.include || args.exclusive) && !args.includedPatches.contains(patchName)) {
+            logger.info("$prefix: Excluded by default")
+            return@patchLoop
         }
 
         logger.trace("Adding $patchName")
@@ -70,7 +70,7 @@ fun Patcher.applyPatchesVerbose() {
 }
 
 fun Patcher.mergeFiles() {
-    this.addFiles(args.patchArgs?.patchingArgs!!.mergeFiles) { file ->
+    this.addIntegrations(args.patchArgs?.patchingArgs!!.mergeFiles) { file ->
         logger.info("Merging $file")
     }
 }